+++ src/lib/project/migration/migrateProjectResources.ts
@@ -8,6 +8,7 @@ import {
   migrate420r1To420r2,
   migrate420r2To420r3,
   migrate420r3To420r4,
+  migrate420r4To420r5,
 } from "./versions/410to420";
 
 const migrations: ProjectResourcesMigration[] = [
@@ -16,6 +17,7 @@ const migrations: ProjectResourcesMigration[] = [
   migrate420r1To420r2,
   migrate420r2To420r3,
   migrate420r3To420r4,
+  migrate420r4To420r5,
 ];
 
 const lastMigration = migrations[migrations.length - 1];